Added the mini login page
This is a simple login page, with only front-end work done. No backend, the fields still need to be linked. 
It has a login button, username field, password field, and forgot password button which ill do things in the future but for now its a basic shell and can be changed any moment.